nvim: refactor markdown zettelkasten setup
The intention behind the refactor is so that if the `zk` binary is unavailable
then we will still try to manually create apply the template settings. A fair
bit of the code is still hardcoded. In the future, it would be nice to parse
`zk/config.toml` and apply it's settings as necessary